Rename the folders and files to python django default
