+++ community/tests/test_license.py
@@ -1,6 +1,6 @@
 from django.test import RequestFactory, TestCase
-from community.Models.LicenseModel import License
-from community.Views.LicenseView import LicenseView
+from community.models.license_model import License
+from community.views.license_view import LicenseView
 from datetime import date
 
 class LicenseTest(TestCase):
@@ -10,7 +10,17 @@ class LicenseTest(TestCase):
             owner='cleber', repo='cremilda', have_license=True, date=date.today()
         )
     
-    def test_details(self):
+    def test_RepositoryExistence(self):
+        request = self.factory.get('/community/license/cleber/cremilda')
+
+        # request.license = self.license
+
+        # response = LicenseView().get(request,owner='cleber',repo='cremilda')
+        # response = LicenseView(request)
+        response = LicenseView.as_view()(request, 'cleber', 'desenho')
+        self.assertEqual(response.status_code, 404)
+
+    def test_LicenseTrue(self):
         request = self.factory.get('/community/license/cleber/cremilda')
 
         # request.license = self.license
