+++ community/urls.py
@@ -1,5 +1,5 @@
 from django.urls import path
-from community.Views.LicenseView import LicenseView
+from community.views.license_view import LicenseView
 
 urlpatterns = [
     path('license/<str:owner>/<str:repo>/', LicenseView.as_view() ),
